+++ .github/workflows/test-e2e.yml
@@ -16,7 +16,7 @@ jobs:
 
     - uses: actions/setup-go@v5
       with:
-        go-version: 1.23.x
+        go-version: 1.24.x
 
     - name: Build cdk docker image
       run: make build-docker
@@ -32,7 +32,7 @@ jobs:
 
   test-e2e:
     runs-on: ubuntu-latest
-    timeout-minutes: 30
+    timeout-minutes: 60
     needs: build-cdk-image
     strategy:
       fail-fast: false
@@ -51,7 +51,7 @@ jobs:
       with:
         repository: 0xPolygon/kurtosis-cdk
         path: kurtosis-cdk
-        ref: arpit/bridge-service
+        ref: main
     
     - name: Install Kurtosis CDK tools
       uses: ./kurtosis-cdk/.github/actions/setup-kurtosis-cdk
