+++ go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/0xPolygon/cdk-data-availability v0.0.13
 	github.com/0xPolygon/cdk-rpc v0.0.0-20241004114257-6c3cb6eebfb6
 	github.com/0xPolygon/zkevm-ethtx-manager v0.2.11
-	github.com/0xPolygonHermez/zkevm-synchronizer-l1 v1.0.8-0.20250507055106-0321b43a1ee2
+	github.com/0xPolygonHermez/zkevm-synchronizer-l1 v1.0.8-0.20250507070913-a33005d489e6
 	github.com/agglayer/aggkit v0.3.0-beta1.0.20250507060341-dbab1bd391dd
 	github.com/ethereum/go-ethereum v1.15.5
 	github.com/iden3/go-iden3-crypto v0.0.17
