+++ l1infotree/tree_test.go
@@ -1,14 +1,22 @@
 package l1infotree_test
 
 import (
+	"context"
 	"encoding/hex"
 	"encoding/json"
+	"fmt"
 	"os"
+	"path"
 	"testing"
 
 	"github.com/0xPolygon/cdk/l1infotree"
+	"github.com/agglayer/aggkit/db"
+	"github.com/agglayer/aggkit/l1infotreesync"
+	"github.com/agglayer/aggkit/l1infotreesync/migrations"
 	"github.com/agglayer/aggkit/log"
 	"github.com/agglayer/aggkit/test/vectors"
+	"github.com/agglayer/aggkit/tree"
+	"github.com/agglayer/aggkit/tree/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
 )
@@ -128,4 +136,98 @@ func TestAddLeaf2(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, testVector.NewRoot, newRoot)
 	}
+}
+
+func TestProofsFromDifferentTrees(t *testing.T) {
+	fmt.Println("aggregator L1InfoTree ===============================================")
+
+	l1Tree, err := l1infotree.NewL1InfoTree(log.WithFields("test"), types.DefaultHeight, [][32]byte{})
+	require.NoError(t, err)
+
+	leaves := createTestLeaves(t, 2)
+
+	aLeaves := make([][32]byte, len(leaves))
+	for i, leaf := range leaves {
+		aLeaves[i] = l1infotree.HashLeafData(
+			leaf.GlobalExitRoot,
+			leaf.PreviousBlockHash,
+			leaf.Timestamp)
+	}
+
+	aggregatorL1InfoTree, aggregatorRoot, err := l1Tree.ComputeMerkleProof(leaves[0].L1InfoTreeIndex, aLeaves)
+	require.NoError(t, err)
+
+	aggregatorProof := types.Proof{}
+	for i, p := range aggregatorL1InfoTree {
+		aggregatorProof[i] = common.BytesToHash(p[:])
+	}
+
+	fmt.Println(aggregatorRoot)
+	fmt.Println(aggregatorProof)
+	fmt.Println("l1 info tree syncer L1InfoTree ===============================================")
+
+	dbPath := path.Join(t.TempDir(), "l1infotreesyncTestProofsFromDifferentTrees.sqlite")
+	require.NoError(t, migrations.RunMigrations(dbPath))
+
+	dbe, err := db.NewSQLiteDB(dbPath)
+	require.NoError(t, err)
+
+	l1InfoTree := tree.NewAppendOnlyTree(dbe, migrations.L1InfoTreePrefix)
+
+	tx, err := db.NewTx(context.Background(), dbe)
+	require.NoError(t, err)
+
+	for _, leaf := range leaves {
+		err = l1InfoTree.AddLeaf(tx, leaf.BlockNumber, leaf.BlockPosition, types.Leaf{
+			Index: leaf.L1InfoTreeIndex,
+			Hash:  leaf.Hash,
+		})
+
+		require.NoError(t, err)
+	}
+
+	require.NoError(t, tx.Commit())
+
+	l1InfoTreeSyncerRoot, err := l1InfoTree.GetRootByIndex(context.Background(), leaves[1].L1InfoTreeIndex)
+	require.NoError(t, err)
+	l1InfoTreeSyncerProof, err := l1InfoTree.GetProof(
+		context.Background(),
+		leaves[0].L1InfoTreeIndex,
+		l1InfoTreeSyncerRoot.Hash,
+	)
+	require.NoError(t, err)
+	for i, l := range aggregatorL1InfoTree {
+		require.Equal(t, common.Hash(l), l1InfoTreeSyncerProof[i])
+	}
+
+	fmt.Println(leaves[0].GlobalExitRoot)
+	fmt.Println(l1InfoTreeSyncerProof)
+
+	require.Equal(t, aggregatorRoot, l1InfoTreeSyncerRoot.Hash)
+	require.Equal(t, aggregatorProof, l1InfoTreeSyncerProof)
+}
+
+func createTestLeaves(t *testing.T, numOfLeaves int) []*l1infotreesync.L1InfoTreeLeaf {
+	t.Helper()
+
+	leaves := make([]*l1infotreesync.L1InfoTreeLeaf, 0, numOfLeaves)
+
+	for i := 0; i < numOfLeaves; i++ {
+		leaf := &l1infotreesync.L1InfoTreeLeaf{
+			L1InfoTreeIndex:   uint32(i),
+			Timestamp:         uint64(i),
+			BlockNumber:       uint64(i),
+			BlockPosition:     uint64(i),
+			PreviousBlockHash: common.HexToHash(fmt.Sprintf("0x%x", i)),
+			MainnetExitRoot:   common.HexToHash(fmt.Sprintf("0x%x", i)),
+			RollupExitRoot:    common.HexToHash(fmt.Sprintf("0x%x", i)),
+		}
+
+		leaf.GlobalExitRoot = leaf.GetGlobalExitRoot()
+		leaf.Hash = leaf.GetHash()
+
+		leaves = append(leaves, leaf)
+	}
+
+	return leaves
 }
