+++ test/combinations/fork11-rollup.yml
@@ -1,7 +1,7 @@
 args:
   zkevm_contracts_image: leovct/zkevm-contracts:v7.0.0-rc.2-fork.11-patch.1
   zkevm_prover_image: hermeznetwork/zkevm-prover:v7.0.2-fork.11
-  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.1.2
+  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.61.19
   zkevm_node_image: hermeznetwork/zkevm-node:v0.7.0-fork11-RC1
   cdk_node_image: cdk
   gas_token_enabled: true
