+++ test/combinations/fork12-cdk-validium.yml
@@ -1,10 +1,10 @@
 args:
   zkevm_contracts_image: leovct/zkevm-contracts:v8.0.0-rc.4-fork.12-patch.1
   zkevm_prover_image: hermeznetwork/zkevm-prover:v8.0.0-RC12-fork.12
-  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.1.2
+  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.61.19
   cdk_node_image: cdk
   gas_token_enabled: true
   gas_token_address: ''
   data_availability_mode: cdk-validium
   sequencer_type: erigon
-  agglayer_image: ghcr.io/agglayer/agglayer@sha256:5715855f2cc6834bd84a99a33937915164648547d21bfb55198948b0ae4e0fad
+  agglayer_image: ghcr.io/agglayer/agglayer:0.3.0-rc.16
