+++ test/combinations/fork9-cdk-validium.yml
@@ -1,7 +1,7 @@
 args:
   zkevm_contracts_image: leovct/zkevm-contracts:v6.0.0-rc.1-fork.9-patch.1
   zkevm_prover_image: hermeznetwork/zkevm-prover:v6.0.6
-  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.1.2
+  cdk_erigon_node_image: hermeznetwork/cdk-erigon:v2.61.19
   zkevm_node_image: hermeznetwork/zkevm-node:v0.7.3-RC1
   cdk_validium_node_image: 0xpolygon/cdk-validium-node:0.7.0-cdk
   cdk_node_image: cdk
